pyproject.toml
[tool.poetry]
name = "suthing"
version = "0.1.0"
description = "Small Useful Things : python package for convenient primitives"
authors = ["Alexander Belikov <abelikov@gmail.com>"]
license = "BSD-3-Clause"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.7.1"
black = "^23.3.0"
isort = "5.11.5"
mypy = "^1.1.1"
pyyaml = "^6.0"
pandas = "1.3.5"
types-pyyaml = "^6.0.12.9"
typing-extensions = "^4.5.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
